Name posters as "poster.jpg"
Original script makes the poster file name based on the movie/show file name which IMO is a bad idea because it is always subject to changes (upgraded movie/tv file ex). Of course this would be better as a command line option to toggle on or off or customize the default name but I didn't need that for my use case. In the article below by Plex, they list a number of default names that posters can have.

https://support.plex.tv/articles/200220677-local-media-assets-movies/